requirements/default.txt
django
django-celery
